@@ -1,65 +1,52 @@
-import React, { Component } from 'react'
-import axios from 'axios'
-import { format } from 'timeago.js'
-import { Link } from 'react-router-dom'
+import { useEffect, useState } from "react";
+import axios from "axios";
+import { Link } from "react-router-dom";
 
-export default class NotesList extends Component {
+export function NotesList() {
+  const [notes, setNotes] = useState([]);
 
-    state = {
-        notes: []
-    }
+  useEffect(() => {
+    const getNotes = async () => {
+      const res = await axios.get("http://localhost:4000/api/notes");
+      console.log(res);
+      setNotes(res.data);
+    };
+    getNotes();
+  }, []);
 
-    async componentDidMount() {
-        this.getNotes();
-    }
+  async function deleteNote(noteId) {
+    const res = await axios.delete("http://localhost:4000/api/notes/" + noteId);
+    if (res.status === 204)
+      setNotes([...notes.filter((note) => note._id !== noteId)]);
+  }
 
-    getNotes = async () => {
-        const res = await axios.get('http://localhost:4000/api/notes')
-        this.setState({
-            notes: res.data
-        });
-    }
-
-    deleteNote = async (noteId) => {
-        await axios.delete('http://localhost:4000/api/notes/' + noteId);
-        this.getNotes();
-    }
-
-    render() {
-        return (
-            <div className="row">
-                {
-                    this.state.notes.map(note => (
-                        <div className="col-md-4 p-2" key={note._id}>
-                            <div className="card">
-                                <div className="card-header d-flex justify-content-between">
-                                    <h5>{note.title}</h5>
-                                    <Link to={"/edit/" + note._id} className="btn btn-secondary">
-                                        <i className="material-icons">
-                                            border_color</i>
-                                    </Link>
-                                </div>
-                                <div className="card-body">
-                                    <p>
-                                        {note.content}
-                                    </p>
-                                    <p>
-                                        Author: {note.author}
-                                    </p>
-                                    <p>
-                                        {format(note.createdAt)}
-                                    </p>
-                                </div>
-                                <div className="card-footer">
-                                    <button className="btn btn-danger" onClick={() => this.deleteNote(note._id)}>
-                                        Delete
-                                    </button>
-                                </div>
-                            </div>
-                        </div>
-                    ))
-                }
+  return (
+    <div className="row">
+      {notes.map((note) => (
+        <div className="col-md-4 p-2" key={note._id}>
+          <div className="card rounded-0">
+            <div className="card-header d-flex justify-content-between align-items-center">
+              <h5>{note.title}</h5>
+              <Link to={"/edit/" + note._id} className="btn btn-sm">
+                <i className="material-icons">border_color</i>
+              </Link>
             </div>
-        )
-    }
+            <div className="card-body">
+              <p>{note.content}</p>
+              <p>Author: {note.author}</p>
+              {/* <p>{format(note.createdAt)}</p> */}
+            </div>
+            <div className="card-footer">
+              <button
+                className="btn btn-danger"
+                onClick={() => deleteNote(note._id)}
+              >
+                Delete
+              </button>
+            </div>
+          </div>
+        </div>
+      ))}
+    </div>
+  );
 }
